refactor : hange directory
