+++ src/app/(root)/mypage/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { LikeBakeryList } from "@/components/bakeries/LikeBakeryList";
+import UserProfile from "@/components/commons/profile/UserProfile";
 import { useUserStore } from "@/store/userStore";
 import Link from "next/link";
-import UserProfile from "@/components/commons/UserProfile";
 
 const MyPage = () => {
   const { nickname, profile, description } = useUserStore((state) => ({
