+++ src/components/bakeries/BakeryCard.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import phoneIcon from "../../../public/image/icons/phone.png";
 import defaultImg from "../../../public/image/noimg.png";
-import LikeButton from "./LikeButton";
+import LikeButton from "../commons/LikeButton";
 
 type BakeryCardProps = {
   bakeryId: string;
