+++ src/components/bakeries/BakeryList.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { useEffect, useState } from "react";
 import { createClient } from "@/supabase/client";
 import { Bakery } from "@/types/bakery";
-import { BakeryCard } from "../commons/BakeryCard";
+import { BakeryCard } from "./BakeryCard";
 import { bakeryItem } from "@/app/api/supabase/bakery/route";
 
 type BakeryListProp = {
