+++ src/components/comment/CommentItem.tsx
@@ -6,7 +6,7 @@ import { useUserStore } from "@/store/userStore";
 import { User } from "@/types/user";
 import { useEffect, useState } from "react";
 import { ToastContainer } from "react-toastify";
-import UserProfile from "../commons/UserProfile";
+import UserProfile from "../commons/profile/UserProfile";
 import { showConfirmToast } from "../commons/toast/ConfirmToast";
 
 type CommentItem = {
