+++ src/components/comment/CommentList.tsx
@@ -5,7 +5,7 @@ import { useUserStore } from "@/store/userStore";
 import { Comment } from "@/types/comment";
 import { ChangeEvent, FormEvent, useEffect, useState } from "react";
 import Toast from "../commons/toast/Toast";
-import UserProfile from "../commons/UserProfile";
+import UserProfile from "../commons/profile/UserProfile";
 import CommentItem from "./CommentItem";
 
 type CommentProps = {
