+++ src/components/map/StoreInformation.tsx
